Formatterクラスを作成し、役割を分けて作成
