--- 05.ls/class
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require 'pathname'
-require 'byebug'
-
-require './argument'
-
-class ShortFormat
-
